pyproject.toml
[tool.poetry]
name = "itcourse"
version = "0.1.0"
description = ""
authors = ["iimgera <iimgera28062004@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.1.7"
django-solo = "^2.0.0"
djangorestframework = "^3.14.0"
django-jazzmin = "^2.6.0"
drf-yasg = "^1.21.5"
cors = "^1.0.1"
django-cors-headers = "^3.14.0"
djangorestframework-simplejwt = "5.2.2"
six = "^1.16.0"
django-utils-six = "^2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
